Added boredless_traveler project from Codecademy Computer Science track
This project helps tourist find attractions in their destination of choice by checking keywords against others and matching attractions to the tourist based on destination and interests. This project used most of the skills learned so far up, but with heavy usage of lists. Seems like they are pretty essential in Python